+++ NEWS.md
@@ -1,9 +1,18 @@
-# mapedit 0.2.1
+# mapedit 0.3.0
+
+### API Changes
+
+* **BREAKING** editFeatures and selectFeatures add a map argument
+instead of platform
+
+### New Features
 
 * add editFeatures function for easy add, edit, delete with existing simple features (sf)
+* add editMap.NULL
 * add record argument to edit* functions to preserve the series
 of actions from an editing session.  If `record = TRUE` then a `"recorder"` attribute will be added to the returned object for
 full reproducibility.
+* add internal playback for recorded session for future use
 * add viewer argument to select and edit functions to allow
 user the flexibility to adjust the viewer experience.  Default
 will be paneViewer() in an attempt to keep the workflow
@@ -11,6 +20,10 @@ within one RStudio window/context.
 * change height to 97% to fill viewer
 * document more thoroughly
 * pass trial CRAN check
+
+## Bug Fixes
+
+* fix internal `combine_list_of_sf` with length 0 `list`; found when `editFeatures()` and save with no changes
 
 # mapedit 0.2.0